+++ setup.sh
@@ -466,9 +466,9 @@ if [[ -d "$SWIFT_UI_DIR" ]]; then
     <key>CFBundleName</key>
     <string>Local Whisper</string>
     <key>CFBundleVersion</key>
-    <string>1.6.11</string>
+    <string>1.6.12</string>
     <key>CFBundleShortVersionString</key>
-    <string>1.6.11</string>
+    <string>1.6.12</string>
     <key>NSPrincipalClass</key>
     <string>NSApplication</string>
     <key>LSUIElement</key>